Minor fixes and added comments for keeper folder.

+++ x/cosmos/keeper/slahsing_event.go
@@ -6,6 +6,13 @@ import (
 	cosmosTypes "github.com/persistenceOne/pstake-native/x/cosmos/types"
 )
 
+/*
+Adds the slashing message entry to the slashing store with the given validator address.
+Performs the following actions :
+  1. Checks if store has the key or not. If not then create new entry
+  2. Checks if store has it and matches all the details present in the message. If not then create a new entry.
+  3. Finally, if all the details match then append the validator address to keep track.
+*/
 func (k Keeper) setSlashingEventDetails(ctx sdk.Context, msg cosmosTypes.MsgSlashingEventOnCosmosChain, validatorAddress sdk.ValAddress) {
 	slashingStore := prefix.NewStore(ctx.KVStore(k.storeKey), cosmosTypes.KeySlashingStore)
 	chainIDHeightAndTxHash := cosmosTypes.NewChainIDHeightAndTxHash(msg.ChainID, msg.BlockHeight, msg.SlashType)
@@ -39,6 +46,19 @@ func (k Keeper) setSlashingEventDetails(ctx sdk.Context, msg cosmosTypes.MsgSlas
 	}
 }
 
+// Sets the addedToCValue flag true for thw given slashing store value
+func (k Keeper) setAddedToCValueTrue(ctx sdk.Context, value cosmosTypes.SlashingStoreValue) {
+	slashingStore := prefix.NewStore(ctx.KVStore(k.storeKey), cosmosTypes.KeySlashingStore)
+	chainIDHeightAndTxHash := cosmosTypes.NewChainIDHeightAndTxHash(value.SlashingDetails.ChainID, value.SlashingDetails.BlockHeight, value.SlashingDetails.SlashType)
+
+	var slashingStoreValue cosmosTypes.SlashingStoreValue
+	k.cdc.MustUnmarshal(slashingStore.Get(k.cdc.MustMarshal(&chainIDHeightAndTxHash)), &slashingStoreValue)
+
+	slashingStoreValue.AddedToCValue = true
+	slashingStore.Set(k.cdc.MustMarshal(&chainIDHeightAndTxHash), k.cdc.MustMarshal(&slashingStoreValue))
+}
+
+// Gets all the slashing event details present in the slashing store
 func (k Keeper) getAllSlashingEventDetails(ctx sdk.Context) (list []cosmosTypes.SlashingStoreValue) {
 	slashingStore := prefix.NewStore(ctx.KVStore(k.storeKey), cosmosTypes.KeySlashingStore)
 	iterator := slashingStore.Iterator(nil, nil)
@@ -51,12 +71,21 @@ func (k Keeper) getAllSlashingEventDetails(ctx sdk.Context) (list []cosmosTypes.
 	return list
 }
 
+// Removes the slashing event details corresponding to the passed values
 func (k Keeper) deleteSlashingEventDetails(ctx sdk.Context, value cosmosTypes.SlashingStoreValue) {
 	slashingStore := prefix.NewStore(ctx.KVStore(k.storeKey), cosmosTypes.KeySlashingStore)
 	chainIDHeightAndTxHash := cosmosTypes.NewChainIDHeightAndTxHash(value.SlashingDetails.ChainID, value.SlashingDetails.BlockHeight, value.SlashingDetails.SlashType)
 	slashingStore.Delete(k.cdc.MustMarshal(&chainIDHeightAndTxHash))
 }
 
+/*
+ProcessAllSlashingEvents processes all the slashing requests
+This function is called every EndBlocker to perform the defined set of actions as mentioned below :
+   1. Get the list of all slashing requests
+   2. Checks if the majority of the validator oracle have sent the minting request. Also checks the addedToCValue flag.
+   3. If majority is reached and other conditions match then slashing event is accepted and C value is updated
+   4. Another condition of ActiveBlockHeight is also checked whether to delete the entry or not.
+*/
 func (k Keeper) ProcessAllSlashingEvents(ctx sdk.Context) {
 	slashingEventList := k.getAllSlashingEventDetails(ctx)
 	for _, se := range slashingEventList {
@@ -72,15 +101,18 @@ func (k Keeper) ProcessAllSlashingEvents(ctx sdk.Context) {
 			slashedAmount := delegation.Sub(se.SlashingDetails.CurrentDelegation)
 			// update C value based on slashed amount
 			k.SlashingEvent(ctx, slashedAmount)
+			// set added to C value true
+			k.setAddedToCValueTrue(ctx, se)
 			// update current delegations of the validator with the delegation supplied with message
 			k.UpdateDelegationCosmosValidator(ctx, valAddress, se.SlashingDetails.CurrentDelegation)
 		}
-		if se.ActiveBlockHeight < ctx.BlockHeight() {
+		if se.ActiveBlockHeight < ctx.BlockHeight() && se.AddedToCValue {
 			k.deleteSlashingEventDetails(ctx, se)
 		}
 	}
 }
 
+// StoreValueEqualOrNotSlashingEvent Helper function for slashing store to check if the relevant details in the message matches or not.
 func StoreValueEqualOrNotSlashingEvent(storeValue cosmosTypes.MsgSlashingEventOnCosmosChain, msgValue cosmosTypes.MsgSlashingEventOnCosmosChain) bool {
 	if storeValue.ValidatorAddress != msgValue.ValidatorAddress {
 		return false
